Added TensorFlow
Model is created, can get fed data and return a prediction.
Reward is based on a raw computation of the best move at any given time (based on score difference before and after pseudo-move).
The game ends if after any move the board is unchanged or no play is available. GameOver gets a hard punishment (since it's the absolute no-no).
The gradient descent doesn't work yet so it's not evolving.